+++ src/test/java/org/jivesoftware/openfire/carbons/MessageCarbonsTest.java
@@ -1,9 +1,10 @@
 package org.jivesoftware.openfire.carbons;
 
-import junit.framework.Assert;
 import org.jivesoftware.openfire.forward.Forwarded;
 import org.junit.Test;
 import org.xmpp.packet.Message;
+
+import static org.junit.Assert.assertEquals;
 
 /**
  * @author Christian Schudt
@@ -20,7 +21,7 @@ public class MessageCarbonsTest {
 
         Sent sent = new Sent(forwarded);
         String xml = sent.getElement().asXML();
-        Assert.assertEquals("<sent xmlns=\"urn:xmpp:carbons:2\"><forwarded xmlns=\"urn:xmpp:forward:0\"><message xmlns=\"jabber:client\" type=\"chat\"><body>Tests</body></message></forwarded></sent>", xml);
+        assertEquals("<sent xmlns=\"urn:xmpp:carbons:2\"><forwarded xmlns=\"urn:xmpp:forward:0\"><message xmlns=\"jabber:client\" type=\"chat\"><body>Tests</body></message></forwarded></sent>", xml);
     }
 
     @Test
@@ -33,6 +34,6 @@ public class MessageCarbonsTest {
 
         Received received = new Received(forwarded);
         String xml = received.getElement().asXML();
-        Assert.assertEquals("<received xmlns=\"urn:xmpp:carbons:2\"><forwarded xmlns=\"urn:xmpp:forward:0\"><message xmlns=\"jabber:client\" type=\"chat\"><body>Tests</body></message></forwarded></received>", xml);
+        assertEquals("<received xmlns=\"urn:xmpp:carbons:2\"><forwarded xmlns=\"urn:xmpp:forward:0\"><message xmlns=\"jabber:client\" type=\"chat\"><body>Tests</body></message></forwarded></received>", xml);
     }
 }
