+++ src/test/java/org/jivesoftware/openfire/forward/ForwardTest.java
@@ -1,8 +1,9 @@
 package org.jivesoftware.openfire.forward;
 
-import junit.framework.Assert;
 import org.junit.Test;
 import org.xmpp.packet.Message;
+
+import static org.junit.Assert.assertEquals;
 
 /**
  * @author Christian Schudt
@@ -17,6 +18,6 @@ public class ForwardTest {
 
         Forwarded forwarded = new Forwarded(message);
         String xml = forwarded.getElement().asXML();
-        Assert.assertEquals("<forwarded xmlns=\"urn:xmpp:forward:0\"><message xmlns=\"jabber:client\" type=\"chat\"><body>Tests</body></message></forwarded>", xml);
+        assertEquals("<forwarded xmlns=\"urn:xmpp:forward:0\"><message xmlns=\"jabber:client\" type=\"chat\"><body>Tests</body></message></forwarded>", xml);
     }
 }
