+++ src/test/java/org/jivesoftware/openfire/handler/IQEntityTimeHandlerTest.java
@@ -1,6 +1,5 @@
 package org.jivesoftware.openfire.handler;
 
-import junit.framework.Assert;
 import org.dom4j.Element;
 import org.jivesoftware.util.XMPPDateTimeFormat;
 import org.junit.Test;
@@ -12,6 +11,9 @@ import java.util.Date;
 import java.util.GregorianCalendar;
 import java.util.TimeZone;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
 /**
  * @author csh
  */
@@ -20,16 +22,16 @@ public class IQEntityTimeHandlerTest {
     @Test
     public void testIQInfo() {
         IQEntityTimeHandler iqEntityTimeHandler = new IQEntityTimeHandler();
-        Assert.assertTrue(iqEntityTimeHandler.getFeatures().hasNext());
-        Assert.assertEquals(iqEntityTimeHandler.getFeatures().next(), "urn:xmpp:time");
-        Assert.assertEquals(iqEntityTimeHandler.getInfo().getNamespace(), "urn:xmpp:time");
-        Assert.assertEquals(iqEntityTimeHandler.getInfo().getName(), "time");
+        assertTrue(iqEntityTimeHandler.getFeatures().hasNext());
+        assertEquals(iqEntityTimeHandler.getFeatures().next(), "urn:xmpp:time");
+        assertEquals(iqEntityTimeHandler.getInfo().getNamespace(), "urn:xmpp:time");
+        assertEquals(iqEntityTimeHandler.getInfo().getName(), "time");
     }
 
     @Test
     public void testTimeZone() {
         IQEntityTimeHandler iqEntityTimeHandler = new IQEntityTimeHandler();
-        Assert.assertEquals(iqEntityTimeHandler.formatsTimeZone(TimeZone.getTimeZone("GMT-8:00")), "-08:00");
+        assertEquals(iqEntityTimeHandler.formatsTimeZone(TimeZone.getTimeZone("GMT-8:00")), "-08:00");
     }
 
     @Test
@@ -39,7 +41,7 @@ public class IQEntityTimeHandlerTest {
         Calendar calendar = new GregorianCalendar();
         calendar.setTime(date);
         calendar.setTimeZone(TimeZone.getTimeZone("GMT"));
-        Assert.assertEquals(iqEntityTimeHandler.getUtcDate(date), DatatypeConverter.printDateTime(calendar));
+        assertEquals(iqEntityTimeHandler.getUtcDate(date), DatatypeConverter.printDateTime(calendar));
     }
 
     @Test
@@ -67,12 +69,12 @@ public class IQEntityTimeHandlerTest {
         IQEntityTimeHandler iqEntityTimeHandler = new IQEntityTimeHandler();
         IQ input = new IQ(IQ.Type.get, "1");
         IQ result = iqEntityTimeHandler.handleIQ(input);
-        Assert.assertEquals(result.getChildElement().getName(), "time");
-        Assert.assertEquals(result.getChildElement().getNamespace().getText(), "urn:xmpp:time");
-        Assert.assertEquals(result.getChildElement().content().size(), 2);
-        Assert.assertTrue(result.getChildElement().content().get(0) instanceof Element);
-        Assert.assertTrue(result.getChildElement().content().get(1) instanceof Element);
-        Assert.assertEquals(((Element) result.getChildElement().content().get(0)).getName(), "tzo");
-        Assert.assertEquals(((Element) result.getChildElement().content().get(1)).getName(), "utc");
+        assertEquals(result.getChildElement().getName(), "time");
+        assertEquals(result.getChildElement().getNamespace().getText(), "urn:xmpp:time");
+        assertEquals(result.getChildElement().content().size(), 2);
+        assertTrue(result.getChildElement().content().get(0) instanceof Element);
+        assertTrue(result.getChildElement().content().get(1) instanceof Element);
+        assertEquals(((Element) result.getChildElement().content().get(0)).getName(), "tzo");
+        assertEquals(((Element) result.getChildElement().content().get(1)).getName(), "utc");
     }
 }
